// src/app/shared/services/certificate-pdf/interface.interface.ts
export interface  pdfInfoSolicitud {
    noControl: string;
    nombre: string;
    telefono: string;
    seguroSocial: string;
    email: string;
    carrera: string;
    domicilio: string;
    noSeguro: string;
    periodoProyect: string;
    nombreProyecto: string;
    opcionelEgida: string;
    noResidentes: string;
    nameEmpresa: string;
    rfc: string;
    giro: string;
    calleNumber: string;
    titularName: string;
    puestotitular: string;
    asesorExt: string;
    puestoAseExt: string;
    colonia: string;
    cp: string;
    ciudadEstado: string;
    telefonoEmpresa: string;
    intermediarioName: string;
    puestoIntermediario: string;
    misionEmpresa: string;
}